Graduate spell.rs
The last transpiled function in the spell subsystem, `ex_spellrepall`,
is rewritten, and the parent takes `#![deny(unsafe_op_in_unsafe_fn)]`
along with the rest of the tree. spell.rs is 368 lines and 2 unsafe: a
module header, the shared constants and globals, and one command.

The constants got names for the types c2rust left anonymous —
`WordFlags`, `SpellResult`, `FindMode`, `CharType` — and the four
families that are only used inside `spell/` are typed to what consumes
them (`c_int`), which retires 34 `as c_int` casts across the submodules.
`matchinf_T`'s fields are documented; its two magic array sizes are
`MAXWLEN` and `MAXWLEN + 1` rather than 254 and 255.

Dead after the carves and deleted: `DUMPFLAG_*` (dump.rs has its own),
`MB_MAXBYTES` (from mbyte.rs now), `NULL`, `true_0`, `false_0`,
`DEFAULT_MAXPATHL`, and the 38-line `pub use types::{...}` re-export
blob, trimmed to the eighteen names still referenced.

`p_ws` is still round-tripped through a `bool`, so a non-zero
'wrapscan' comes back as 1 — the C behaviour, and unobservable while it
is a boolean option, but not worth silently changing.

Ratchet shrinks with no growth to allow: spell.rs 865 -> 368 lines,
8 -> 2 unsafe, `files_without_deny_unsafe_op` 225 -> 224, clippy
12,534 -> 12,529 in one fewer file.

Verified against the pre-batch binary: spellsweep byte- and
report-identical, the `:spelldump` probe identical, 86 spell oldtest
executions, 26 functional spell tests plus buffer_updates_spec (the only
`:spellrepall` coverage in the tree), plain and under ASan.

+++ crates/nvim/src/nvim/spell/lang.rs
@@ -392,10 +392,10 @@ pub unsafe fn parse_spelllang(wp: *mut win_T) -> *mut c_char {
                     strcasecmp(lang.as_ptr(), (*slang).sl_name) == 0
                 };
                 if matches {
-                    let mut region_mask = REGION_ALL as c_int;
+                    let mut region_mask = REGION_ALL;
                     if !filename && !region.is_null() {
                         let c = find_region((*slang).sl_regions.as_ptr(), region);
-                        if c == REGION_ALL as c_int {
+                        if c == REGION_ALL {
                             if (*slang).sl_add {
                                 if (*slang).sl_regions[0] != 0 {
                                     // This addition file covers other regions.
@@ -512,10 +512,10 @@ pub unsafe fn parse_spelllang(wp: *mut win_T) -> *mut c_char {
                 }
 
                 if !slang.is_null() {
-                    let mut region_mask = REGION_ALL as c_int;
+                    let mut region_mask = REGION_ALL;
                     if !use_region.is_null() && !dont_use_region {
                         let c = find_region((*slang).sl_regions.as_ptr(), use_region);
-                        if c != REGION_ALL as c_int {
+                        if c != REGION_ALL {
                             region_mask = 1 << c;
                         } else if (*slang).sl_regions[0] != 0 {
                             // This spell file covers other regions.
@@ -603,7 +603,7 @@ unsafe fn find_region(rp: *const c_char, region: *const c_char) -> c_int {
         let mut i = 0;
         loop {
             if *rp.offset(i as isize) == 0 {
-                return REGION_ALL as c_int;
+                return REGION_ALL;
             }
             if *rp.offset(i as isize) == *region && *rp.offset(i as isize + 1) == *region.offset(1)
             {
